Use cwrp instead of promise chain and this.setState
This entire logic goes inside the reducer that handles the email check
action:

	// default state, or initial state when there's no action
	const newState = { emailExists: true };

	if (response.data.external_id) {
		newState = {
		showPasswordField: true,
		flow: FLOW.LOGIN,
		emailExists: false,
		};
	} else {
		newState = {
		showPasswordField: true,
		flow: FLOW.SIGNUP,
		emailExists: false,
		};
	}
	});

	return newState;

This value gets propagated as a prop to the component. We use a
`componentWillReceiveProps` because now the life-cycle of the response
becomes: action triggered -> response came back -> reducer returns new
state -> this gets passed to component.

The first benefit is that the last step happens only if the component is
mounted. In the old promise callback style, we need to explicitly check
if the component is mounted inside each callback because without that,
`this.setState` call will fail, or for that matter any `this.<function>`

Inside `componentWillReceiveProps`, we check if the data actually
changed. We can do this using a simple function like shallowEqual (which
is present in underscore, and is around 5 lines of code and a
generally nice to have util). We update the state only if the data
actually changed. This is the second advantage. We don't trigger
re-renders when it's not required. In the old case, we have to add these
state comparison checks inside every callback in the promise chain.